Reranking test
